feat: Updated autollm/utils/document_reading.py

+++ autollm/utils/document_reading.py
@@ -101,7 +101,7 @@ def read_github_repo_as_documents(
     """
 
     # Ensure the temp_dir directory exists
-    temp_dir = Path("autollm/temp/")
+    temp_dir = Path("./autollm/temp/")
     temp_dir.mkdir(parents=True, exist_ok=True)
 
     logger.info(f"Cloning github repo {git_repo_url} into temporary directory {temp_dir}..")
@@ -119,7 +119,7 @@ def read_github_repo_as_documents(
         logger.info(f"Operations complete, deleting temporary directory {temp_dir}..")
     finally:
         # Delete the temporary directory
-        shutil.rmtree(temp_dir, onerror=on_rm_error)
+        shutil.rmtree(temp_dir, onerror=on_rm_error) if temp_dir.is_dir() else logger.warning("Temporary directory does not exist.")
 
     return documents
